requirements.txt
Django==5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
psycopg2==2.9.9
tensorflow-cpu==2.19.0  # Use CPU-only version
h5py==3.13.0
numpy==1.26.4
Pillow==10.3.0
mysqlclient==2.2.7
python-decouple==3.8